GF-1418 - fixes bug in the frontend javascript wordcount.
It looks like the frontend was getting confused by whitespace, I'll
add my test case to the jira ticket.  I've altered the approach so
that it will now simply split on whitespace and then delete any
empty string elements in the array.  We still need to count 0 for
example as a word in it's own right (when seperated by whitespace).