[1459] When saving the data from the form we were using the CostItemValidator to convert from CostItemValues to a CostItem, but it was the CostItemValues that were being stored in the database and subsequently re-loaded from the section for display, with validation re-applied. But the value of `cost` stored contained the comma and when we tried to convert that to a BigDecimal directly it failed (the CostItemValidator normalises the string to remove commas).
The fix in this commit causes the validated value, i.e. the CostItem, to be converted to json and stored in the database. This ties item handling even more closely to cost items, and away from a more generalised mechanism, but we'll live with that for the moment.

I think it points towards a more general solution not only for items, but for all section answer documents where we can update FieldCheck to return not just errors, but a ValidatedNel[FieldError, JsObject] and, if valid, we can store that object in the database instead of the raw field strings. We'd also need a mechansim to convert back from the stronger types to the strings to be sent to display in a form. This is a fair bit of work and I haven't tried to address that at this point.