Quite a bit of re-working to unify the handling of the cost list with the other sections. I'm now treating the list as a Field and rendering it similarly to the other fields. This has meant needing to pass more information into the renderers than was previously needed, namely the application overview and the form section (so that it can render the links in the table appropriately).
The other big change is that I'm no longer flattening the json objects that represent the answers for a section in `ActionHandler`. Instead, I'm passing the `JsObject` to the various functions that render pages. In many cases the `Field` implementation flattens the json before passing the answers into its rendering template, but in the case of `CostListField` it needs access to the array of items and `JsArray`s are not handled well by `flatten`, hence the advantage of working with the `JsObject` at that point.

These changes have touched a lot of files, but mostly it's simply changes to method signatures propagating through.